Move Gc into an internal module
