+++ gc/src/gc.rs
@@ -1,211 +1,255 @@
-use std::ptr;
-use std::mem;
-use std::cell::{Cell, RefCell};
-use trace::Trace;
+//! Thread-local garbage-collected boxes (The `Gc<T>` type).
+//!
+//! The `Gc<T>` type provides shared ownership of an immutable value.
+//! It is marked as non-sendable because the garbage collection only occurs
+//! thread locally.
 
-// XXX Obviously not 100 bytes GC threshold - choose a number
-const GC_THRESHOLD: usize = 100;
+use std::cell::{self, Cell, RefCell, BorrowState};
+use std::ops::{Deref, DerefMut, CoerceUnsized};
+use std::marker;
+use gc_internals::GcBox;
+use trace::{Trace, Tracer};
 
-struct GcState {
-    bytes_allocated: usize,
-    boxes_start: Option<Box<GcBoxTrait + 'static>>,
-    boxes_end: *mut Option<Box<GcBoxTrait + 'static>>,
+// We expose the force_collect method from the gc internals
+pub use gc_internals::force_collect;
+
+////////
+// Gc //
+////////
+
+/// A garbage-collected pointer type over an immutable value.
+///
+/// See the [module level documentation](./) for more details.
+pub struct Gc<T: Trace + ?Sized + 'static> {
+    // XXX We can probably take advantage of alignment to store this
+    root: Cell<bool>,
+    _ptr: *mut GcBox<T>,
 }
 
-/// Whether or not the thread is currently in the sweep phase of garbage collection
-/// During this phase, attempts to dereference a Gc<T> pointer will trigger a panic
-thread_local!(static GC_SWEEPING: Cell<bool> = Cell::new(false));
+impl<T: ?Sized> !marker::Send for Gc<T> {}
 
-/// The garbage collector's internal state.
-thread_local!(static GC_STATE: RefCell<GcState> = RefCell::new(GcState {
-    bytes_allocated: 0,
-    boxes_start: None,
-    boxes_end: ptr::null_mut(),
-}));
+impl<T: ?Sized> !marker::Sync for Gc<T> {}
 
-pub struct GcBoxHeader {
-    // XXX This is horribly space inefficient - not sure if we care
-    // We are using a word word bool - there is a full 63 bits of unused data :(
-    roots: Cell<usize>,
-    next: Option<Box<GcBoxTrait + 'static>>,
-    marked: Cell<bool>,
-}
+impl<T: Trace + ?Sized + marker::Unsize<U>, U: Trace + ?Sized> CoerceUnsized<Gc<U>> for Gc<T> {}
 
-/// Internal trait - must be implemented by every garbage collected allocation
-/// GcBoxTraits form a linked list of allocations.
-trait GcBoxTrait {
-    /// Get a reference to the internal GcBoxHeader
-    fn header(&self) -> &GcBoxHeader;
-
-    /// Get a mutable reference to the internal GcBoxHeader
-    fn header_mut(&mut self) -> &mut GcBoxHeader;
-
-    /// Initiate a marking trace through the GcBoxTrait
-    unsafe fn mark_value(&self);
-
-    /// Get the size of the allocationr required to create the GcBox
-    fn size_of(&self) -> usize;
-}
-
-pub struct GcBox<T: Trace + ?Sized + 'static> {
-    header: GcBoxHeader,
-    data: T,
-}
-
-impl<T: Trace> GcBox<T> {
-    /// Allocate a garbage collected GcBox on the heap,
-    /// and append it to the thread local GcBox chain.
+impl<T: Trace> Gc<T> {
+    /// Constructs a new `Gc<T>`.
     ///
-    /// The GcBox allocated this way starts it's life
-    /// rooted.
-    pub fn new(value: T) -> *mut GcBox<T> {
-        GC_STATE.with(|_st| {
-            let mut st = _st.borrow_mut();
+    /// # Collection
+    ///
+    /// This method could trigger a Garbage Collection.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use gc::Gc;
+    ///
+    /// let five = Gc::new(5);
+    /// ```
+    pub fn new(value: T) -> Gc<T> {
+        unsafe {
+            // Allocate the memory for the object
+            let ptr = GcBox::new(value);
 
-            // XXX We should probably be more clever about collecting
-            if st.bytes_allocated > GC_THRESHOLD {
-                collect_garbage(&mut *st);
-            }
-
-            let mut gcbox = Box::new(GcBox {
-                header: GcBoxHeader {
-                    roots: Cell::new(1),
-                    marked: Cell::new(false),
-                    next: None,
-                },
-                data: value,
-            });
-
-            let gcbox_ptr = &mut *gcbox as *mut _;
-
-            let next_boxes_end = &mut gcbox.header.next as *mut _;
-            if st.boxes_end.is_null() {
-                assert!(st.boxes_start.is_none(),
-                        "If something had been allocated, boxes_end would be set");
-                // The next place we're going to add something!
-                st.boxes_end = next_boxes_end;
-                st.boxes_start = Some(gcbox);
-            } else {
-                unsafe {
-                    *st.boxes_end = Some(gcbox);
-                }
-                st.boxes_end = next_boxes_end;
-            }
-
-            // We allocated some bytes! Let's record it
-            st.bytes_allocated += mem::size_of::<GcBox<T>>();
-
-            // Return the pointer to the newly allocated data
-            gcbox_ptr
-        })
+            // When we create a Gc<T>, all pointers which have been moved to the
+            // heap no longer need to be rooted, so we unroot them.
+            (*ptr).value()._gc_unroot();
+            Gc { _ptr: ptr, root: Cell::new(true) }
+        }
     }
 }
 
-impl<T: Trace + ?Sized> GcBox<T> {
-    /// Mark this GcBox, and trace through it's data
-    pub unsafe fn mark(&self) {
-        let marked = self.header.marked.get();
-        if !marked {
-            self.header.marked.set(true);
-            self.data._gc_mark();
+impl<T: Trace + ?Sized> Gc<T> {
+    #[inline]
+    fn inner(&self) -> &GcBox<T> {
+        unsafe { &*self._ptr }
+    }
+}
+
+impl<T: Trace + ?Sized> Trace for Gc<T> {
+    #[inline]
+    unsafe fn _trace<U: Tracer>(&self, _: U) { /* do nothing */ }
+
+    #[inline]
+    unsafe fn _gc_mark(&self) {
+        self.inner().mark();
+    }
+
+    #[inline]
+    unsafe fn _gc_root(&self) {
+        assert!(!self.root.get(), "Can't double-root a Gc<T>");
+        self.root.set(true);
+
+        self.inner().root();
+    }
+
+    #[inline]
+    unsafe fn _gc_unroot(&self) {
+        assert!(self.root.get(), "Can't double-unroot a Gc<T>");
+        self.root.set(false);
+
+        self.inner().unroot();
+    }
+}
+
+impl<T: Trace + ?Sized> Clone for Gc<T> {
+    #[inline]
+    fn clone(&self) -> Gc<T> {
+        unsafe { self.inner().root(); }
+        Gc { _ptr: self._ptr, root: Cell::new(true) }
+    }
+}
+
+impl<T: Trace + ?Sized> Deref for Gc<T> {
+    type Target = T;
+
+    #[inline]
+    fn deref(&self) -> &T {
+        &self.inner().value()
+    }
+}
+
+impl<T: Trace + ?Sized> Drop for Gc<T> {
+    #[inline]
+    fn drop(&mut self) {
+        // If this pointer was a root, we should unroot it.
+        if self.root.get() {
+            unsafe { self.inner().unroot(); }
+        }
+    }
+}
+
+////////////
+// GcCell //
+////////////
+
+/// A mutable memory location with dynamically checked borrow rules
+/// which can be used inside of a garbage collected pointer.
+///
+/// This object is a RefCell which can be used inside of a Gc<T>.
+pub struct GcCell<T: ?Sized + 'static> {
+    rooted: Cell<bool>,
+    cell: RefCell<T>,
+}
+
+impl <T: Trace> GcCell<T> {
+    /// Creates a new `GcCell` containing `value`.
+    #[inline]
+    pub fn new(value: T) -> GcCell<T> {
+        GcCell {
+            rooted: Cell::new(true),
+            cell: RefCell::new(value),
         }
     }
 
-    /// Increase the root count on this GcBox.
-    /// Roots prevent the GcBox from being destroyed by
-    /// the garbage collector.
-    pub unsafe fn root(&self) {
-        self.header.roots.set(self.header.roots.get() + 1);
-    }
-
-    /// Decrease the root count on this GcBox.
-    /// Roots prevent the GcBox from being destroyed by
-    /// the garbage collector.
-    pub unsafe fn unroot(&self) {
-        self.header.roots.set(self.header.roots.get() - 1);
-    }
-
-    /// Get the value form the GcBox
-    pub fn value(&self) -> &T {
-        // XXX This may be too expensive, but will help catch errors with
-        // accessing Gc values in destructors.
-        GC_SWEEPING.with(|sweeping| assert!(!sweeping.get(),
-                                            "Gc pointers may be invalid when GC is running"));
-        &self.data
+    /// Consumes the `GcCell`, returning the wrapped value.
+    #[inline]
+    pub fn into_inner(self) -> T {
+        self.cell.into_inner()
     }
 }
 
-impl<T: Trace> GcBoxTrait for GcBox<T> {
-    fn header(&self) -> &GcBoxHeader { &self.header }
-
-    fn header_mut(&mut self) -> &mut GcBoxHeader { &mut self.header }
-
-    unsafe fn mark_value(&self) { self.mark() }
-
-    fn size_of(&self) -> usize { mem::size_of::<T>() }
-}
-
-/// Collect some garbage!
-fn collect_garbage(st: &mut GcState) {
-    let mut next_node = &mut st.boxes_start
-        as *mut Option<Box<GcBoxTrait + 'static>>;
-
-    // Mark
-    loop {
-        if let Some(ref mut node) = *unsafe { &mut *next_node } {
-            {
-                // XXX This virtual method call is nasty :(
-                let header = node.header_mut();
-                next_node = &mut header.next as *mut _;
-
-                // If it doesn't have roots - we can abort now
-                if header.roots.get() == 0 { continue }
-            }
-            // We trace in a different scope such that node isn't
-            // mutably borrowed anymore
-            unsafe { node.mark_value(); }
-        } else { break }
+impl <T: Trace + ?Sized> GcCell<T> {
+    /// Immutably borrows the wrapped value.
+    ///
+    /// The borrow lasts until the returned `GcCellRef` exits scope.
+    /// Multiple immutable borrows can be taken out at the same time.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the value is currently mutably borrowed.
+    #[inline]
+    pub fn borrow(&self) -> GcCellRef<T> {
+        self.cell.borrow()
     }
 
-    GC_SWEEPING.with(|collecting| collecting.set(true));
+    /// Mutably borrows the wrapped value.
+    ///
+    /// The borrow lasts until the returned `GcCellRefMut` exits scope.
+    /// The value cannot be borrowed while this borrow is active.
+    ///
+    /// #Panics
+    ///
+    /// Panics if the value is currently borrowed.
+    #[inline]
+    pub fn borrow_mut(&self) -> GcCellRefMut<T> {
+        let val_ref = self.cell.borrow_mut();
 
-    let mut next_node = &mut st.boxes_start
-        as *mut Option<Box<GcBoxTrait + 'static>>;
+        // Force the val_ref's contents to be rooted for the duration of the mutable borrow
+        if !self.rooted.get() {
+            unsafe { val_ref._gc_root(); }
+        }
 
-    // Sweep
-    loop {
-        if let Some(ref mut node) = *unsafe { &mut *next_node } {
-            // XXX This virtual method call is nasty :(
-            let size = node.size_of();
-            let header = node.header_mut();
-
-            if header.marked.get() {
-                // This node has already been marked - we're done!
-                header.marked.set(false);
-                next_node = &mut header.next;
-            } else {
-                // The node wasn't marked - we need to delete it
-                st.bytes_allocated -= size;
-                let mut tmp = None;
-                mem::swap(&mut tmp, &mut header.next);
-                mem::swap(&mut tmp, unsafe { &mut *next_node });
-
-                // At this point, the node is destroyed if it exists due to tmp dropping
-            }
-        } else { break }
+        GcCellRefMut {
+            _ref: val_ref,
+            _rooted: &self.rooted,
+        }
     }
-
-    // Update the end pointer to point to the correct location
-    st.boxes_end = next_node;
-
-    // XXX This should probably be done with some kind of finally guard
-    GC_SWEEPING.with(|collecting| collecting.set(false));
 }
 
-/// Immediately trigger a garbage collection on the current thread.
-pub fn force_collect() {
-    GC_STATE.with(|_st| {
-        let mut st = _st.borrow_mut();
-        collect_garbage(&mut *st);
-    });
+impl<T: Trace + ?Sized> Trace for GcCell<T> {
+    #[inline]
+    unsafe fn _trace<U: Tracer>(&self, _: U) { /* do nothing */ }
+
+    #[inline]
+    unsafe fn _gc_mark(&self) {
+        match self.cell.borrow_state() {
+            BorrowState::Writing => (),
+            _ => self.cell.borrow()._gc_mark(),
+        }
+    }
+
+    #[inline]
+    unsafe fn _gc_root(&self) {
+        assert!(!self.rooted.get(), "Can't root a GcCell Twice!");
+        self.rooted.set(true);
+
+        match self.cell.borrow_state() {
+            BorrowState::Writing => (),
+            _ => self.cell.borrow()._gc_root(),
+        }
+    }
+
+    #[inline]
+    unsafe fn _gc_unroot(&self) {
+        assert!(self.rooted.get(), "Can't unroot a GcCell Twice!");
+        self.rooted.set(false);
+
+        match self.cell.borrow_state() {
+            BorrowState::Writing => (),
+            _ => self.cell.borrow()._gc_unroot(),
+        }
+    }
+}
+
+/// A wrapper type for an immutably borrowed value from a GcCell<T>
+pub type GcCellRef<'a, T> = cell::Ref<'a, T>;
+
+/// A wrapper type for a mutably borrowed value from a GcCell<T>
+pub struct GcCellRefMut<'a, T: Trace + ?Sized + 'static> {
+    _ref: ::std::cell::RefMut<'a, T>,
+    _rooted: &'a Cell<bool>,
+}
+
+impl<'a, T: Trace + ?Sized> Deref for GcCellRefMut<'a, T> {
+    type Target = T;
+
+    #[inline]
+    fn deref(&self) -> &T { &*self._ref }
+}
+
+impl<'a, T: Trace + ?Sized> DerefMut for GcCellRefMut<'a, T> {
+    #[inline]
+    fn deref_mut(&mut self) -> &mut T { &mut *self._ref }
+}
+
+impl<'a, T: Trace + ?Sized> Drop for GcCellRefMut<'a, T> {
+    #[inline]
+    fn drop(&mut self) {
+        // Restore the rooted state of the GcCell's contents to the state of the GcCell.
+        // During the lifetime of the GcCellRefMut, the GcCell's contents are rooted.
+        if !self._rooted.get() {
+            unsafe { self._ref._gc_unroot(); }
+        }
+    }
 }
