+++ gc/tests/gc_semantics.rs
@@ -5,7 +5,7 @@ extern crate gc;
 
 use std::cell::Cell;
 use std::thread::LocalKey;
-use gc::{Trace, GcCell, Gc, force_collect};
+use gc::{Trace, GcCell, Gc, gc_force_collect};
 
 // Utility methods for the tests
 #[derive(PartialEq, Eq, Debug, Clone, Copy)]
@@ -90,12 +90,12 @@ fn basic_allocate() {
     {
         let _gced_val = Gc::new(GcWatch(&FLAGS));
         FLAGS.with(|f| assert_eq!(f.get(), GcWatchFlags::new(0, 0, 1, 0)));
-        force_collect();
+        gc_force_collect();
         FLAGS.with(|f| assert_eq!(f.get(), GcWatchFlags::new(1, 0, 1, 0)));
     }
 
     FLAGS.with(|f| assert_eq!(f.get(), GcWatchFlags::new(1, 0, 1, 0)));
-    force_collect();
+    gc_force_collect();
     FLAGS.with(|f| assert_eq!(f.get(), GcWatchFlags::new(1, 0, 1, 1)));
 }
 
@@ -119,7 +119,7 @@ fn basic_cycle_allocate() {
         FLAGS1.with(|f| assert_eq!(f.get(), GcWatchFlags::new(0, 0, 1, 0)));
         FLAGS2.with(|f| assert_eq!(f.get(), GcWatchFlags::new(0, 0, 1, 0)));
 
-        force_collect();
+        gc_force_collect();
 
         FLAGS1.with(|f| assert_eq!(f.get(), GcWatchFlags::new(1, 0, 1, 0)));
         FLAGS2.with(|f| assert_eq!(f.get(), GcWatchFlags::new(1, 0, 1, 0)));
@@ -131,7 +131,7 @@ fn basic_cycle_allocate() {
             FLAGS1.with(|f| assert_eq!(f.get(), GcWatchFlags::new(1, 0, 1, 0)));
             FLAGS2.with(|f| assert_eq!(f.get(), GcWatchFlags::new(1, 0, 1, 0)));
 
-            force_collect();
+            gc_force_collect();
 
             FLAGS1.with(|f| assert_eq!(f.get(), GcWatchFlags::new(2, 0, 1, 0)));
             FLAGS2.with(|f| assert_eq!(f.get(), GcWatchFlags::new(2, 0, 1, 0)));
@@ -140,7 +140,7 @@ fn basic_cycle_allocate() {
         FLAGS1.with(|f| assert_eq!(f.get(), GcWatchFlags::new(2, 0, 1, 0)));
         FLAGS2.with(|f| assert_eq!(f.get(), GcWatchFlags::new(2, 0, 1, 0)));
 
-        force_collect();
+        gc_force_collect();
 
         FLAGS1.with(|f| assert_eq!(f.get(), GcWatchFlags::new(3, 0, 1, 0)));
         FLAGS2.with(|f| assert_eq!(f.get(), GcWatchFlags::new(3, 0, 1, 0)));
@@ -149,7 +149,7 @@ fn basic_cycle_allocate() {
     FLAGS1.with(|f| assert_eq!(f.get(), GcWatchFlags::new(3, 0, 1, 0)));
     FLAGS2.with(|f| assert_eq!(f.get(), GcWatchFlags::new(3, 0, 1, 0)));
 
-    force_collect();
+    gc_force_collect();
 
     FLAGS1.with(|f| assert_eq!(f.get(), GcWatchFlags::new(3, 0, 1, 1)));
     FLAGS2.with(|f| assert_eq!(f.get(), GcWatchFlags::new(3, 0, 1, 1)));
@@ -189,7 +189,7 @@ fn gccell_rooting() {
         FLAGS.with(|f| assert_eq!(f.get(), GcWatchFlags::new(0, 0, 1, 0)));
 
         // It should be traced by the GC
-        force_collect();
+        gc_force_collect();
         FLAGS.with(|f| assert_eq!(f.get(), GcWatchFlags::new(1, 0, 1, 0)));
 
         {
@@ -202,7 +202,7 @@ fn gccell_rooting() {
             FLAGS.with(|f| assert_eq!(f.get(), GcWatchFlags::new(1, 0, 1, 0)));
 
             // It should be traced by the GC
-            force_collect();
+            gc_force_collect();
             FLAGS.with(|f| assert_eq!(f.get(), GcWatchFlags::new(2, 0, 1, 0)));
         }
 
@@ -215,7 +215,7 @@ fn gccell_rooting() {
 
             // It shouldn't be traced by the GC (as it's owned by the GcCell)
             // If it had rootable members, they would be traced by the GC
-            force_collect();
+            gc_force_collect();
             FLAGS.with(|f| assert_eq!(f.get(), GcWatchFlags::new(2, 1, 1, 0)));
         }
 
@@ -223,12 +223,12 @@ fn gccell_rooting() {
         FLAGS.with(|f| assert_eq!(f.get(), GcWatchFlags::new(2, 1, 2, 0)));
 
         // It should be traced by the GC
-        force_collect();
+        gc_force_collect();
         FLAGS.with(|f| assert_eq!(f.get(), GcWatchFlags::new(3, 1, 2, 0)))
     }
 
     // It should be collected by the GC
-    force_collect();
+    gc_force_collect();
     FLAGS.with(|f| assert_eq!(f.get(), GcWatchFlags::new(3, 1, 2, 1)))
 }
